requirements.txt
ansicolors==1.0.2
argh==0.26.1
args==0.1.0
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
clint==0.4.1
configparser==3.5.0b2
coverage==4.0.3
Django==1.8.7
django-appconf==1.0.1
django-compressor==2.0
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-environ==0.4.0
django-grappelli==2.7.3
django-gulp==2.1.1
django-mailgun==0.8.0
django-modelcluster==1.1
django-nose==1.4.2
django-overextends==0.4.1
django-taggit==0.18.0
django-treebeard==3.0
djangorestframework==3.3.2
env-tools==2.0.0
factory-boy==2.6.0
fake-factory==0.5.3
flake8==2.5.1
futures==2.2.0
gunicorn==19.4.1
html5lib==0.9999999
iso8601==0.1.11
mccabe==0.3.1
nose==1.3.7
pathtools==0.1.2
pep8==1.5.7
Pillow==3.0.0
polytester==1.2.0
psutil==2.2.1
psycopg2==2.6.1
pyflakes==1.0.0
pytz==2015.7
PyYAML==3.11
rcssmin==1.0.6
recurly==2.2.17
requests==2.9.1
rjsmin==1.0.12
six==1.10.0
sphinx-me==0.3
sqlparse==0.1.18
tini==3.0.1
Unidecode==0.4.19
wagtail==1.3.1
wagtailmodeladmin==2.2.6
watchdog==0.8.2
wheel==0.24.0
whitenoise==2.0.6
Willow==0.2.2
